Design Update 2.0 : Login Page finalized
BIG UPDATES HERE, so beware

1st, I put some (weird) changes to some add-on that didn't work as intended. If needed, we'll erase this push and only add the design-related files.

otherwise, I've redone my logistic, changed a few files to ease the comprehension.
Basically, the Login page is updated.
The color thing was scratched in favor of a simpler Light/Dark theme.

What remains is to have a clean logo, and merge it with the Hard-Code with the database and stuff.

ALSO, I left a class called "PleaseDontTouch" where some currently unused methods are located, which I may need down the line.

enjoy !!